+++ mobile/android/android-components/docs/changelog.md
@@ -21,6 +21,7 @@ permalink: /changelog/
 * **feature-tabs**
   * Adds `lastAccess` to the `Tab` data class that is used in `TabsTray`.
   * Adds a new SelectOrAddUseCase for reopening existing tab with matching HistoryMetadataKey. [#10611](https://github.com/mozilla-mobile/android-components/issues/10611)
+  * Adds `recoverable` parameter to `RemoveAllTabsUseCase.invoke()` to specify whether `UndoMiddleware` should make the removed tabs recoverable.
 
 # 91.0.0
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v91.0.0...master)